Refactor All Trips and Active Trip tabs to have their own stacks. Data persists when switching btw different tabs now
