+++ drawerScreens/DrawerNavigatorRoutes.tsx
@@ -1,27 +1,17 @@
-import React, { useState } from "react";
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View, SafeAreaView, Image } from "react-native"
-import { createStackNavigator } from "@react-navigation/stack";
+import React, { useContext } from "react";
+import { Image } from "react-native"
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-
-//import the bottom tabs
-import AllTrips from "./bottomNavigation/AllTrips";
-import HomeScreen from "../components/HomeScreen/HomeScreen";
-import ActiveTrip from "./bottomNavigation/ActiveTrip";
 import Login from "./Login";
-// import Archive from "./bottomNavigation/Archive";
 import Archive from "../components/Archive/Archive";
-import { ScreenStackHeaderBackButtonImage } from "react-native-screens";
 import { AuthContext } from "../AuthProvider";
+
+import ActiveTrip from "./ActiveTrip";
+
 
 const Bottom = createBottomTabNavigator();
 
-
 export default function DrawerNavigatorRoutes() {
-  const { username } = React.useContext(AuthContext);
-
-
-
+  // const { username } = useContext(AuthContext);
 
   return (
     <Bottom.Navigator
@@ -66,9 +56,5 @@ export default function DrawerNavigatorRoutes() {
       name="Archive" children={() => (<Archive /> )} />
 
     </Bottom.Navigator>
-
   )
-
 }
-
-// options={{ headerTransparent: true}}
