--- drawerScreens/LoginScreen.tsx
@@ -1,4 +1,0 @@
-import React, { useState } from "react";
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View, SafeAreaView } from "react-native"
-
