--- drawerScreens/Sample2.tsx
@@ -1,13 +1,0 @@
-import React, { useState } from "react";
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View, SafeAreaView } from "react-native"
-
-export default function Sample2 ( ) {
-
-
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>This is another stack example</Text>
-    </View>
-  )
-}
